promote jax-cache to default jax recipe
The jax-cache recipe is now close to a free win for any gradient-based
workflow that goes through tesseract-jax's custom_vjp (which is the
typical path), and incurs only minor overhead on tiny models. Replace
the plain jax template and the vectoradd_jax example with their cached
versions; the now-redundant jax-cache template and vectoradd_jax_cache
example are removed. JSON inputs/test cases in examples/vectoradd_jax/
are unchanged and continue to pass.
