--- examples/vectoradd_jax_cache/checkgradients_example_inputs.json
@@ -1,12 +1,0 @@
-{
-  "inputs": {
-    "a": {
-      "v": [1.0, 2.0, 3.0],
-      "s": 2.0
-    },
-    "b": {
-      "v": [4.0, 5.0, 6.0],
-      "s": 1.0
-    }
-  }
-}
