--- examples/vectoradd_jax_cache/example_abstract_inputs.json
@@ -1,12 +1,0 @@
-{
-  "inputs": {
-    "a": {
-      "v": { "shape": [3], "dtype": "float32" },
-      "s": { "shape": [], "dtype": "float32" }
-    },
-    "b": {
-      "v": { "shape": [3], "dtype": "float32" },
-      "s": { "shape": [], "dtype": "float32" }
-    }
-  }
-}
