--- examples/vectoradd_jax_cache/example_bad_inputs.json
@@ -1,11 +1,0 @@
-{
-  "inputs": {
-    "a": {
-      "v": [1.0, 2.0, 3.0]
-    },
-    "b": {
-      "v": [4.0, 5.0, 6.0],
-      "x": 3.0
-    }
-  }
-}
