--- examples/vectoradd_jax_cache/example_bad_inputs2.json
@@ -1,19 +1,0 @@
-{
-  "inputs": {
-    "a": {
-      "v": [1.0, 2.0, 3.0]
-    },
-    "b": {
-      "v": [4.0, 5.0, 6.0],
-      "s": {
-        "object_type": "array",
-        "shape": [1],
-        "dtype": "float32",
-        "data": {
-          "buffer": [1],
-          "encoding": "json"
-        }
-      }
-    }
-  }
-}
