--- examples/vectoradd_jax_cache/example_bad_vjp_inputs3.json
@@ -1,14 +1,0 @@
-{
-  "inputs": {
-    "a": {
-      "v": [1.0, 2.0, 3.0]
-    },
-    "b": {
-      "v": [4.0, 5.0, 6.0],
-      "s": 3.0
-    }
-  },
-  "vjp_inputs": ["a.v", "b.v"],
-  "vjp_outputs": [],
-  "cotangent_vector": {}
-}
