--- examples/vectoradd_jax_cache/example_bad_vjp_inputs5.json
@@ -1,13 +1,0 @@
-{
-  "inputs": {
-    "a": {
-      "v": [1.0, 2.0, 3.0]
-    },
-    "b": {
-      "v": [4.0, 5.0, 6.0],
-      "s": 3.0
-    }
-  },
-  "vjp_inputs": [],
-  "cotangent_vector": {}
-}
