--- examples/vectoradd_jax_cache/example_jacobian_inputs.json
@@ -1,13 +1,0 @@
-{
-  "inputs": {
-    "a": {
-      "v": [1.0, 2.0, 3.0]
-    },
-    "b": {
-      "v": [4.0, 5.0, 6.0],
-      "s": 3.0
-    }
-  },
-  "jac_inputs": ["a.v", "a.s"],
-  "jac_outputs": ["vector_add.result", "vector_add.normed_result"]
-}
