--- examples/vectoradd_jax_cache/example_vjp_inputs.json
@@ -1,17 +1,0 @@
-{
-  "inputs": {
-    "a": {
-      "v": [1.0, 2.0, 3.0]
-    },
-    "b": {
-      "v": [4.0, 5.0, 6.0],
-      "s": 3.0
-    }
-  },
-  "vjp_inputs": ["a.v", "b.v"],
-  "vjp_outputs": ["vector_add.result", "vector_add.normed_result"],
-  "cotangent_vector": {
-    "vector_add.result": [0.1, 0.2, 0.3],
-    "vector_add.normed_result": [0.4, 0.5, 0.6]
-  }
-}
