--- examples/vectoradd_jax_cache/example_vjp_inputs2.json
@@ -1,16 +1,0 @@
-{
-  "inputs": {
-    "a": {
-      "v": [1.0, 2.0, 3.0]
-    },
-    "b": {
-      "v": [4.0, 5.0, 6.0],
-      "s": 3.0
-    }
-  },
-  "vjp_inputs": ["a.v"],
-  "vjp_outputs": ["vector_add.result"],
-  "cotangent_vector": {
-    "vector_add.result": [0.1, 0.2, 0.3]
-  }
-}
