--- examples/vectoradd_jax_cache/tesseract_requirements.txt
@@ -1,2 +1,0 @@
-jax[cpu]
-equinox
