--- examples/vectoradd_jax_cache/test_cases/test_abstract_eval.json
@@ -1,53 +1,0 @@
-{
-  "endpoint": "abstract_eval",
-  "expected_outputs": {
-    "vector_add": {
-      "result": {
-        "shape": [3],
-        "dtype": "float32"
-      },
-      "normed_result": {
-        "shape": [3],
-        "dtype": "float32"
-      }
-    },
-    "vector_min": {
-      "result": {
-        "shape": [3],
-        "dtype": "float32"
-      },
-      "normed_result": {
-        "shape": [3],
-        "dtype": "float32"
-      }
-    }
-  },
-  "expected_exception": null,
-  "expected_exception_regex": null,
-  "atol": 1e-8,
-  "rtol": 1e-5,
-  "payload": {
-    "inputs": {
-      "a": {
-        "v": {
-          "shape": [3],
-          "dtype": "float32"
-        },
-        "s": {
-          "shape": [],
-          "dtype": "float32"
-        }
-      },
-      "b": {
-        "v": {
-          "shape": [3],
-          "dtype": "float32"
-        },
-        "s": {
-          "shape": [],
-          "dtype": "float32"
-        }
-      }
-    }
-  }
-}
