--- examples/vectoradd_jax_cache/test_cases/test_vjp_wrt_vector.json
@@ -1,59 +1,0 @@
-{
-  "endpoint": "vector_jacobian_product",
-  "expected_outputs": {
-    "a.v": {
-      "object_type": "array",
-      "shape": [3],
-      "dtype": "float32",
-      "data": {
-        "buffer": "zcxMPs3MzD6amRk/",
-        "encoding": "base64"
-      }
-    }
-  },
-  "expected_exception": null,
-  "expected_exception_regex": null,
-  "atol": 1e-8,
-  "rtol": 1e-5,
-  "payload": {
-    "inputs": {
-      "a": {
-        "v": {
-          "object_type": "array",
-          "shape": [3],
-          "dtype": "float64",
-          "data": {
-            "buffer": "AAAAAAAA8D8AAAAAAAAAQAAAAAAAAAhA",
-            "encoding": "base64"
-          }
-        },
-        "s": 2
-      },
-      "b": {
-        "v": {
-          "object_type": "array",
-          "shape": [3],
-          "dtype": "float64",
-          "data": {
-            "buffer": "AAAAAAAAEEAAAAAAAAAUQAAAAAAAABhA",
-            "encoding": "base64"
-          }
-        },
-        "s": 1
-      }
-    },
-    "vjp_inputs": ["a.v"],
-    "vjp_outputs": ["vector_add.result"],
-    "cotangent_vector": {
-      "vector_add.result": {
-        "object_type": "array",
-        "shape": [3],
-        "dtype": "float64",
-        "data": {
-          "buffer": "mpmZmZmZuT+amZmZmZnJPzMzMzMzM9M/",
-          "encoding": "base64"
-        }
-      }
-    }
-  }
-}
