--- examples/vectoradd_jax_cache/test_cases_inputs/example_checkgradients_input.json
@@ -1,28 +1,0 @@
-{
-  "inputs": {
-    "a": {
-      "v": {
-        "object_type": "array",
-        "shape": [3],
-        "dtype": "int64",
-        "data": {
-          "buffer": "AQAAAAAAAAACAAAAAAAAAAMAAAAAAAAA",
-          "encoding": "base64"
-        }
-      },
-      "s": 2
-    },
-    "b": {
-      "v": {
-        "object_type": "array",
-        "shape": [3],
-        "dtype": "int64",
-        "data": {
-          "buffer": "BAAAAAAAAAAFAAAAAAAAAAYAAAAAAAAA",
-          "encoding": "base64"
-        }
-      },
-      "s": 1
-    }
-  }
-}
