--- tesseract_core/sdk/templates/jax-cache/tesseract_api.py
@@ -1,256 +1,0 @@
-# Copyright 2025 Pasteur Labs. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-# Tesseract API module for {{name}}
-# Generated by tesseract {{version}} on {{timestamp}}
-
-from typing import Any
-
-import equinox as eqx
-import jax
-import jax.numpy as jnp
-from pydantic import BaseModel
-
-from tesseract_core.runtime import Differentiable, Float32
-from tesseract_core.runtime.tree_transforms import (
-    LRUCache,
-    filter_func,
-    flatten_with_paths,
-    hash_pytree_leaves,
-    set_at_path,
-)
-
-#
-# Cache configuration
-#
-# When apply() is called, this template runs jax.vjp internally and caches the
-# resulting backward function. A subsequent vector_jacobian_product call on the
-# same inputs reuses that cached backward, skipping the redundant forward pass
-# that vjp would otherwise repeat. Typical speedup is ~10-20%
-# for a val-and-grad on moderate-to-deep networks.
-#
-# The cache also helps when a single apply is followed by multiple vjp calls
-# on the same input — for example, jax.jacrev (which decomposes into one vjp
-# per output basis vector) or CG-style inverse-problem solvers (each inner CG
-# step computes J^T u at the same iterate). Each subsequent vjp on the same
-# input reuses the cached residuals.
-#
-# Poor fit:
-#   - Very small models where the forward pass is microseconds: the cache
-#     machinery overhead exceeds the saved work.
-#
-# Set _CACHE_SIZE = 0 to disable caching entirely; apply() and vjp() then
-# bypass the cache machinery and run their forward passes directly.
-# Increase _CACHE_SIZE for workloads that interleave several apply() calls
-# before their corresponding vjp() calls.
-#
-_CACHE_SIZE = 1
-
-#
-# Schemata
-#
-
-
-# Note: This template uses equinox filter_jit to automatically treat non-array
-# inputs/outputs as static. As Tesseract scalar objects (e.g. Float32) are
-# essentially just wrappers around numpy 0D arrays, they will be considered to
-# be dynamic and will be traced by JAX.
-# If you want to treat scalar numerical values as static you will need to use
-# built-in Python types (e.g. float, int) instead of Float32.
-
-
-class InputSchema(BaseModel):
-    example: Differentiable[Float32]
-
-
-class OutputSchema(BaseModel):
-    example: Differentiable[Float32]
-
-
-#
-# Required endpoints
-#
-
-
-# TODO: Add or import your function here, must be JAX-jittable and
-# take/return a single pytree as an input/output conforming respectively
-# to Input/OutputSchema
-@eqx.filter_jit
-def apply_jit(inputs: dict) -> dict:
-    return inputs
-
-
-def apply(inputs: InputSchema) -> OutputSchema:
-    # Optional: Insert any pre-processing/setup that doesn't require tracing
-    # and is only required when specifically running your apply function
-    # and not your differentiable endpoints.
-    # For example, you might want to set up a logger or mlflow server.
-    # Pre-processing should not modify any input that could impact the
-    # differentiable outputs in a nonlinear way (a constant shift
-    # should be safe)
-
-    inputs_dict = inputs.model_dump()
-
-    if _CACHE_SIZE <= 0:
-        return apply_jit(inputs_dict)
-
-    # Compute forward pass via jax.vjp to cache residuals for a potential
-    # subsequent vector_jacobian_product call. eqx.partition separates
-    # array (differentiable) from non-array outputs; has_aux tells jax.vjp
-    # to only differentiate through the array outputs.
-    def _apply_for_vjp(inputs_dict):
-        out = apply_jit(inputs_dict)
-        diff_out, static_out = eqx.partition(out, eqx.is_array)
-        return diff_out, static_out
-
-    diff_primals, vjp_func, static_primals = jax.vjp(
-        _apply_for_vjp, inputs_dict, has_aux=True
-    )
-    out = eqx.combine(diff_primals, static_primals)
-
-    cotangent_template = jax.tree.map(jnp.zeros_like, diff_primals)
-    _vjp_cache.put(_hash_inputs(inputs_dict), (vjp_func, cotangent_template))
-
-    # Optional: Insert any post-processing that doesn't require tracing
-    # For example, you might want to save to disk or modify a non-differentiable
-    # output. Again, do not modify any differentiable output in a non-linear way.
-    return out
-
-
-#
-# Jax-handled gradient endpoints (no need to modify)
-#
-_vjp_cache = LRUCache(maxsize=_CACHE_SIZE)
-
-
-def _hash_inputs(inputs_dict: dict) -> bytes:
-    """Compute a SHA-256 hash of a pytree of inputs for cache key comparison."""
-    leaves, treedef = jax.tree.flatten(inputs_dict)
-    return hash_pytree_leaves(leaves, treedef)
-
-
-def jacobian(
-    inputs: InputSchema,
-    jac_inputs: set[str],
-    jac_outputs: set[str],
-):
-    return jac_jit(inputs.model_dump(), tuple(jac_inputs), tuple(jac_outputs))
-
-
-def jacobian_vector_product(
-    inputs: InputSchema,
-    jvp_inputs: set[str],
-    jvp_outputs: set[str],
-    tangent_vector: dict[str, Any],
-):
-    return jvp_jit(
-        inputs.model_dump(),
-        tuple(jvp_inputs),
-        tuple(jvp_outputs),
-        tangent_vector,
-    )
-
-
-def vector_jacobian_product(
-    inputs: InputSchema,
-    vjp_inputs: set[str],
-    vjp_outputs: set[str],
-    cotangent_vector: dict[str, Any],
-):
-    inputs_dict = inputs.model_dump()
-
-    # Use get (not pop) so the cached residuals can serve multiple sequential
-    # vjp calls on the same inputs — for example, when tesseract-jax's
-    # value_and_grad is followed by jax.jacrev, which decomposes into many
-    # vjp calls per output basis vector.
-    if (
-        _CACHE_SIZE > 0
-        and (cached := _vjp_cache.get(_hash_inputs(inputs_dict))) is not None
-    ):
-        vjp_func, cotangent_template = cached
-
-        # Build full cotangent from template, zeroing outputs not in vjp_outputs
-        full_cotangent = jax.tree.map(jnp.zeros_like, cotangent_template)
-        full_cotangent = set_at_path(full_cotangent, cotangent_vector)
-
-        (all_input_cotangents,) = vjp_func(full_cotangent)
-        return flatten_with_paths(all_input_cotangents, include_paths=vjp_inputs)
-
-    # Cache disabled or cache miss: fall back to original JIT-compiled path
-    return vjp_jit(
-        inputs_dict,
-        tuple(vjp_inputs),
-        tuple(vjp_outputs),
-        cotangent_vector,
-    )
-
-
-def abstract_eval(abstract_inputs):
-    """Calculate output shape of apply from the shape of its inputs."""
-    is_shapedtype_dict = lambda x: type(x) is dict and (x.keys() == {"shape", "dtype"})
-    is_shapedtype_struct = lambda x: isinstance(x, jax.ShapeDtypeStruct)
-
-    jaxified_inputs = jax.tree.map(
-        lambda x: jax.ShapeDtypeStruct(**x) if is_shapedtype_dict(x) else x,
-        abstract_inputs.model_dump(),
-        is_leaf=is_shapedtype_dict,
-    )
-    dynamic_inputs, static_inputs = eqx.partition(
-        jaxified_inputs, filter_spec=is_shapedtype_struct
-    )
-
-    def wrapped_apply(dynamic_inputs):
-        inputs = eqx.combine(static_inputs, dynamic_inputs)
-        return apply_jit(inputs)
-
-    jax_shapes = jax.eval_shape(wrapped_apply, dynamic_inputs)
-    return jax.tree.map(
-        lambda x: (
-            {"shape": x.shape, "dtype": str(x.dtype)} if is_shapedtype_struct(x) else x
-        ),
-        jax_shapes,
-        is_leaf=is_shapedtype_struct,
-    )
-
-
-#
-# Helper functions
-#
-
-
-@eqx.filter_jit
-def jac_jit(
-    inputs: dict,
-    jac_inputs: tuple[str],
-    jac_outputs: tuple[str],
-):
-    filtered_apply = filter_func(apply_jit, inputs, jac_outputs)
-    return jax.jacrev(filtered_apply)(
-        flatten_with_paths(inputs, include_paths=jac_inputs)
-    )
-
-
-@eqx.filter_jit
-def jvp_jit(
-    inputs: dict, jvp_inputs: tuple[str], jvp_outputs: tuple[str], tangent_vector: dict
-):
-    filtered_apply = filter_func(apply_jit, inputs, jvp_outputs)
-    return jax.jvp(
-        filtered_apply,
-        [flatten_with_paths(inputs, include_paths=jvp_inputs)],
-        [tangent_vector],
-    )[1]
-
-
-@eqx.filter_jit
-def vjp_jit(
-    inputs: dict,
-    vjp_inputs: tuple[str],
-    vjp_outputs: tuple[str],
-    cotangent_vector: dict,
-):
-    filtered_apply = filter_func(apply_jit, inputs, vjp_outputs)
-    _, vjp_func = jax.vjp(
-        filtered_apply, flatten_with_paths(inputs, include_paths=vjp_inputs)
-    )
-    return vjp_func(cotangent_vector)[0]
