--- tesseract_core/sdk/templates/jax-cache/tesseract_config.yaml
@@ -1,31 +1,0 @@
-# Tesseract configuration file
-# Generated by tesseract {{version}} on {{timestamp}}
-
-name: "{{name}}"
-version: "0.1.0"
-description: ""
-
-# Arbitrary user-defined metadata
-# metadata:
-#   key: value
-
-build_config:
-  # Base image to use for the container, must be Ubuntu or Debian-based
-  # base_image: "debian:bookworm-slim"
-
-  # Platform to build the container for. In general, images can only be executed
-  # on the platform they were built for.
-  target_platform: "native"
-
-  # Additional packages to install in the container (via apt-get)
-  # extra_packages:
-  #  - package_name
-
-  # Data to copy into the container, relative to the project root
-  # package_data:
-  #  - [path/to/source, path/to/destination]
-
-  # Additional Dockerfile commands to run during the build process
-  # custom_build_steps:
-  #  - |
-  #   RUN echo "Hello, World!"
