--- tesseract_core/sdk/templates/jax-cache/tesseract_requirements.txt
@@ -1,10 +1,0 @@
-# Tesseract requirements file
-# Generated by tesseract {{version}} on {{timestamp}}
-
-# Add Python requirements like this:
-jax[cpu]
-equinox
-
-# This may contain private dependencies via SSH URLs:
-# git+ssh://git@github.com/username/repo.git@branch
-# (use `tesseract build --forward-ssh-agent` to grant the builder access to your SSH keys)
